requirements.txt
mailersend
grpcio>=1.56.0
grpcio-tools>=1.56.0
grpcio-reflection>=1.56.0
pytest==8.4.2